Refactor: Separa ted types and data transfer objects for role
